part 4 complete

+++ osa4/controllers/blogs.js
@@ -1,26 +1,52 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
-blogsRouter.get('/' || '/api/blogs', async (request, response) => {
-  const blogs = await Blog.find({})
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', {username: 1, name: 1})
+
   response.json(blogs)
 })
-  
+
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id,
+  })
 
   if (blog.title && blog.url) {
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog.id)
+    await user.save()
+
     response.status(201).json(savedBlog)
   } else {
     response.status(400).end()
   }
-
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  let blog = null
+  try {
+    blog = await Blog.findById(request.params.id)
+  } catch(error) {
+    return response.status(404).json({error: 'blog is not found or the id is wrong'})
+  }
+
+  const blogUsersId = blog.user.toString()
+  const tokensId = request.user.id.toString()
+  if (blogUsersId === tokensId) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'you are not authorized to delete blogs of other users' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -30,7 +56,7 @@ blogsRouter.put('/:id', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
